Trait rendy::graph::render::RenderPass

pub trait RenderPass<B, T>: 'static + Send + Sync + Debug where
    B: Backend,
    T: ?Sized, {
    fn name() -> &'static str;

    fn load_shader_sets(
        storage: &'a mut Vec<<B as Backend>::ShaderModule>, 
        factory: &mut Factory<B>, 
        aux: &mut T
    ) -> Vec<GraphicsShaderSet<'a, B>>;

    fn build<impl AsRef<[B::DescriptorSetLayout]>>(
        factory: &mut Factory<B>, 
        aux: &mut T, 
        buffers: &mut [NodeBuffer<'a, B>], 
        images: &mut [NodeImage<'a, B>], 
        sets: &[impl AsRef<[B::DescriptorSetLayout]>]
    ) -> Self
    where
        impl AsRef<[B::DescriptorSetLayout]>: AsRef<[<B as Backend>::DescriptorSetLayout]>;

    fn draw(
        &mut self, 
        layouts: &[<B as Backend>::PipelineLayout], 
        pipelines: &[<B as Backend>::GraphicsPipeline], 
        encoder: RenderPassInlineEncoder<B>, 
        index: usize, 
        aux: &T
    );

    fn dispose(self, factory: &mut Factory<B>, aux: &mut T);

    fn buffers() -> Vec<BufferAccess> { ... }

    fn images() -> Vec<ImageAccess> { ... }

    fn colors() -> usize { ... }

    fn depth() -> bool { ... }

    fn layouts() -> Vec<Layout> { ... }

    fn vertices() -> Vec<(Vec<Element<Format>>, u32)> { ... }

    fn pipelines() -> Vec<Pipeline> { ... }

    fn builder() -> NodeBuilder<B, T> { ... }

    fn prepare(&mut self, _factory: &mut Factory<B>, _aux: &T) -> bool { ... }
}

Render pass.

Required Methods

fn name() -> &'static str

Pass name.

fn load_shader_sets(
storage: &'a mut Vec<<B as Backend>::ShaderModule>,
factory: &mut Factory<B>,
aux: &mut T
) -> Vec<GraphicsShaderSet<'a, B>>

Load shader set. This function should create required shader modules and fill GraphicsShaderSet structure.

Parameters

storage - vector where this function can store loaded modules to give them required lifetime.

factory - Device<B> implementation. B::Device or wrapper.

aux - auxiliary data container. May be anything the implementation desires.

fn build<impl AsRef<[B::DescriptorSetLayout]>>(
factory: &mut Factory<B>,
aux: &mut T,
buffers: &mut [NodeBuffer<'a, B>],
images: &mut [NodeImage<'a, B>],
sets: &[impl AsRef<[B::DescriptorSetLayout]>]
) -> Self where
impl AsRef<[B::DescriptorSetLayout]>: AsRef<[<B as Backend>::DescriptorSetLayout]>,

Build pass instance.

fn draw(
&mut self,
layouts: &[<B as Backend>::PipelineLayout],
pipelines: &[<B as Backend>::GraphicsPipeline],
encoder: RenderPassInlineEncoder<B>,
index: usize,
aux: &T
)

Record drawing commands to the command buffer provided.

fn dispose(self, factory: &mut Factory<B>, aux: &mut T)

Dispose of the pass.

Provided Methods

fn buffers() -> Vec<BufferAccess>

Get set or buffer resources the node uses.

fn images() -> Vec<ImageAccess>

Get set or image resources the node uses.

fn colors() -> usize

Number of color output images.

fn depth() -> bool

Is depth image used.

fn layouts() -> Vec<Layout>

Pipeline layouts

fn vertices() -> Vec<(Vec<Element<Format>>, u32)>

Get vertex input.

fn pipelines() -> Vec<Pipeline>

Graphics pipelines

fn builder() -> NodeBuilder<B, T>

Create NodeBuilder for this node.

fn prepare(&mut self, _factory: &mut Factory<B>, _aux: &T) -> bool

Prepare to record drawing commands.

Should return true if commands must be re-recorded.