Module rendy::resource::buffer

Buffer usage, creation-info and wrappers.

Structs

Buffer
Generic buffer object wrapper.

DownloadBuffer
Type that specify that buffer is intended to be used as staging buffer for data downloads.

IndexBuffer
Type that specify that buffer is intended to be used as index buffer. It implies TRANSFER_DST because device-local, host-invisible memory should be used and transfer is left the only way to fill the buffer.

Info
Buffer info.

UniformBuffer
Type that specify that buffer is intended to be used as uniform buffer. Host visible memory required and device-local preferred.

UploadBuffer
Type that specify that buffer is intended to be used as staging buffer for data uploads.

VertexBuffer
Type that specify that buffer is intended to be used as vertex buffer. It implies TRANSFER_DST because device-local, host-invisible memory should be used and transfer is left the only way to fill the buffer.

Traits

Usage
Usage trait that must implemented by usage types. This trait provides a way to convert type-level usage to the value-level flags.